observe /pyproject.toml
[tool.poetry]
name = "observe "
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langfuse = "^1.0.18"
langchain = "^0.0.286"
python-dotenv = "^1.0.0"
backoff = "^2.2.1"
openai = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
